+++ src/dialect/keywords.rs
@@ -198,6 +198,7 @@ define_keywords!(
     EXECUTE,
     EXISTS,
     EXP,
+    EXPLAIN,
     EXTENDED,
     EXTERNAL,
     EXTRACT,
@@ -472,6 +473,7 @@ define_keywords!(
     VARYING,
     VAR_POP,
     VAR_SAMP,
+    VERBOSE,
     VERSIONING,
     VIEW,
     VIRTUAL,
@@ -494,6 +496,8 @@ define_keywords!(
 pub const RESERVED_FOR_TABLE_ALIAS: &[Keyword] = &[
     // Reserved as both a table and a column alias:
     Keyword::WITH,
+    Keyword::EXPLAIN,
+    Keyword::ANALYZE,
     Keyword::SELECT,
     Keyword::WHERE,
     Keyword::GROUP,
@@ -530,6 +534,8 @@ pub const RESERVED_FOR_TABLE_ALIAS: &[Keyword] = &[
 pub const RESERVED_FOR_COLUMN_ALIAS: &[Keyword] = &[
     // Reserved as both a table and a column alias:
     Keyword::WITH,
+    Keyword::EXPLAIN,
+    Keyword::ANALYZE,
     Keyword::SELECT,
     Keyword::WHERE,
     Keyword::GROUP,